# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample service for managing orders",
        "title": "Orders API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "soberkoder@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:4567",
    "basePath": "/api/v1",
    "paths": {
        "/orders": {
            "get": {
                "description": "Get details of all orders",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get details of all orders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Model.Order"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new order with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "description": "Create order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order Created !",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Some Error Accrue !",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Some Error Accrue !",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/orders/download": {
            "post": {
                "description": "download file on download folder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get details of download file",
                "responses": {
                    "200": {
                        "description": "Downloaded File Done  !",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/orders/files": {
            "get": {
                "description": "Get details of all files of one order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get details of all files of one order",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Model.Data"
                            }
                        }
                    }
                }
            }
        },
        "/orders/id": {
            "get": {
                "description": "Get details of one order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get details of one order",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Model.Order"
                            }
                        }
                    },
                    "400": {
                        "description": "Some Error Accrue !",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Some Error Accrue !",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update on order data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Update order",
                "parameters": [
                    {
                        "description": "Update order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order Updated !",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Some Error Accrue !",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Some Error Accrue !",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete on order by order id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Delete order",
                "parameters": [
                    {
                        "description": "Delete order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Model.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order Deleted !",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Some Error Accrue !",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Some Error Accrue !",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/orders/upload/id": {
            "post": {
                "description": "upload file on folder of order_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get details of upload file",
                "responses": {
                    "200": {
                        "description": "Successfully Uploaded",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Model.Comments": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                }
            }
        },
        "Model.Data": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "Model.ItemList": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "qty": {
                    "type": "integer"
                },
                "totalPrice": {
                    "type": "number"
                },
                "unitPrice": {
                    "type": "number"
                }
            }
        },
        "Model.Order": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Model.Comments"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "itemList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Model.ItemList"
                    }
                },
                "shippingLife": {
                    "$ref": "#/definitions/Model.ShippingLife"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "totalAmount": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "Model.ShippingLife": {
            "type": "object",
            "properties": {
                "shippingMethod": {
                    "type": "string"
                },
                "trackingNumber": {
                    "type": "string"
                }
            }
        }
    }
}